+++ front-end/src/components/Layout/Sidebar.js
@@ -536,7 +536,7 @@ const Sidebar = () => {
             flexShrink: 0,
           }} />
           <Typography variant="caption" color={SIDEBAR_TEXT_DIM} noWrap sx={{ fontWeight: 500 }}>
-            ezeeflo ERP v2.0
+            ezeeflo ERP v3.6.0
           </Typography>
         </Box>
       )}